// apps/avto-shop/src/libs/enums/car.enum.ts
export enum CarStatus {
    ACTIVE = 'ACTIVE',
    SOLD = 'SOLD',
    DELETE = 'DELETE',
}